add excerise 4 folder and file
